feat(Item): separate extensions inside separate classes
