+++ Assets/Sacados/Extensions/ItemExtensions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e1a44bb0328df524591cfb7d5c9ad311
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
